Organize fragment packages
