+++ app/src/main/java/com/rockthevote/grommet/ui/UiModule.java
@@ -7,13 +7,13 @@ import com.rockthevote.grommet.ui.eventFlow.EventPartnerLogin;
 import com.rockthevote.grommet.ui.eventFlow.SessionProgressDialogFragment;
 import com.rockthevote.grommet.ui.eventFlow.SessionSummary;
 import com.rockthevote.grommet.ui.eventFlow.SessionTimeTracking;
-import com.rockthevote.grommet.ui.registration.AdditionalInfoFragment;
-import com.rockthevote.grommet.ui.registration.AssistantInfoFragment;
-import com.rockthevote.grommet.ui.registration.NewRegistrantFragment;
-import com.rockthevote.grommet.ui.registration.PersonalInfoFragment;
+import com.rockthevote.grommet.ui.registration.personal.AdditionalInfoFragment;
+import com.rockthevote.grommet.ui.registration.assistance.AssistantInfoFragment;
+import com.rockthevote.grommet.ui.registration.name.NewRegistrantFragment;
+import com.rockthevote.grommet.ui.registration.address.PersonalInfoFragment;
 import com.rockthevote.grommet.ui.registration.RegistrationActivity;
 import com.rockthevote.grommet.ui.registration.RegistrationCompleteDialogFragment;
-import com.rockthevote.grommet.ui.registration.ReviewAndConfirmFragment;
+import com.rockthevote.grommet.ui.registration.review.ReviewAndConfirmFragment;
 import com.rockthevote.grommet.ui.views.AddressView;
 import com.rockthevote.grommet.ui.views.NameView;
