+++ app/src/main/java/com/rockthevote/grommet/ui/registration/RegistrationPagerAdapter.java
@@ -9,6 +9,11 @@ import android.util.SparseArray;
 import android.view.ViewGroup;
 
 import com.rockthevote.grommet.R;
+import com.rockthevote.grommet.ui.registration.assistance.AssistantInfoFragment;
+import com.rockthevote.grommet.ui.registration.name.NewRegistrantFragment;
+import com.rockthevote.grommet.ui.registration.personal.AdditionalInfoFragment;
+import com.rockthevote.grommet.ui.registration.address.PersonalInfoFragment;
+import com.rockthevote.grommet.ui.registration.review.ReviewAndConfirmFragment;
 
 public class RegistrationPagerAdapter extends FragmentPagerAdapter {
